requirements.txt
uvicorn
fastapi
httpx
python-dotenv
streamlit